+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.7.0] - 2020-05-06
+
+### Added
+* **SharpDPAPI** project
+    * Landed @leftp's `certificates` and `machinecerts` commands
+    * Added `certificates` and `machinecerts` entries to the README.md
+    * Added certificate triage to the `triage` and `machinetriage` commands
+    * Using /password:X now causes the DPAPI masterkey cache to be output
+*  **SharpChrome** project:
+    * Using /password:X now causes the DPAPI masterkey cache to be output
+
+
 ## [1.6.1] - 2020-03-29
 
 ### Changed